requirements.txt
Django==1.10
gunicorn==19.7.1
olefile==0.44
Pillow==4.1.1
whitenoise==3.3.0
dj-database-url==0.4.2
psycopg2==2.7.1
django-rest-framework-social-oauth2==1.0.6
django-bootstrap3==8.2.3
